pyproject.toml
[project]
name = "django-inertia-vuejs"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-extensions>=4.1",
    "django-vite>=3.1.0",
    "inertia-django>=1.2.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]

[tool.ruff.format]
quote-style = "single"
